Add mirroring capabilities
