+++ setup.py
@@ -43,6 +43,7 @@ setup(name="pytroll-file-utils",
                'bin/remove_it.py',
                ],
       data_files=[],
+      packages=['pytroll_fu'],
       zip_safe=False,
       install_requires=['pyinotify', 'posttroll',
                         'trollsift', 'netifaces',
